activities work on a basic premise
