+++ Olympsis-Watchkit/Views/Running/RunActivityPreperation.swift
@@ -14,7 +14,7 @@ struct RunActivityPreperation: View {
     @State private var showLiveActivity: Bool = false
     
     @Environment(\.dismiss) private var dismiss
-    @EnvironmentObject private var manager: ActivityManager
+    @EnvironmentObject private var manager: WorkoutManager
     
     @AppStorage("run_type") private var runType: String?
     
@@ -70,5 +70,5 @@ struct RunActivityPreperation: View {
 
 #Preview {
     RunActivityPreperation()
-        .environmentObject(ActivityManager())
+        .environmentObject(WorkoutManager())
 }
