+++ Olympsis/Views/Activity/CyclingComputerView.swift
@@ -1,0 +1,18 @@
+//
+//  CyclingComputerView.swift
+//  Olympsis
+//
+//  Created by Joel Joseph on 7/23/24.
+//
+
+import SwiftUI
+
+struct CyclingComputerView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    CyclingComputerView()
+}
