+++ Olympsis/Views/Activity/GeneralSportMetrics.swift
@@ -7,12 +7,12 @@
 
 import SwiftUI
 
-struct SoccerMetrics: View {
+struct GeneralSportMetrics: View {
     var body: some View {
         Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
     }
 }
 
 #Preview {
-    SoccerMetrics()
+    GeneralSportMetrics()
 }
