+++ Olympsis/Views/ViewContainer.swift
@@ -101,9 +101,6 @@ struct ViewContainer: View {
             handleRoute(route)
         })
         .task {
-            await session.workoutManager.requestHealthStoreAuthorization()
-            session.workoutManager.loadWeeklyRunHistory()
-            
             session.state = .loading
             await session.CheckIn()
             guard let user = session.user else {
